refactor: moved initClient to helper module
